% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_player_cards.R
\name{build_percentile_geom}
\alias{build_percentile_geom}
\title{build_percentile_geom take the player grade percentiles and build a geom to display ratings boxes}
\usage{
build_percentile_geom(player_row)
}
\arguments{
\item{player_row}{row of data corresponding to the player}
}
\description{
build_percentile_geom take the player grade percentiles and build a geom to display ratings boxes
}
