% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_player_cards.R
\name{build_player_bio}
\alias{build_player_bio}
\title{build_player_bio function to create a ggplot object for a players bio information}
\usage{
build_player_bio(player_row)
}
\arguments{
\item{player_row}{record in total data with players information}
}
\description{
build_player_bio function to create a ggplot object for a players bio information
}
