% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_player_cards.R
\name{build_player_card}
\alias{build_player_card}
\title{build_player_card function to create plots for players}
\usage{
build_player_card(df, player_id)
}
\arguments{
\item{df}{all data}

\item{player_id}{id of player}
}
\description{
build_player_card function to create plots for players
}
