% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_player_cards.R
\name{build_player_radar}
\alias{build_player_radar}
\title{build_player_radar}
\usage{
build_player_radar(df, player_id)
}
\arguments{
\item{df}{entire dataframe}

\item{player_id}{player to return radar for}
}
\description{
build_player_radar
}
