% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_player_cards.R
\name{build_rating_box}
\alias{build_rating_box}
\title{build_rating_box function to create ggplot object given a grade}
\usage{
build_rating_box(rating, label)
}
\arguments{
\item{rating}{scaled percentile grade from 0-100}

\item{label}{string of the box label}
}
\description{
build_rating_box function to create ggplot object given a grade
}
