% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_player_cards.R
\name{load_player_card_data}
\alias{load_player_card_data}
\title{load_player_card_data function to build data used in cards}
\usage{
load_player_card_data()
}
\description{
load_player_card_data function to build data used in cards
}
